# src/cognitive_shield_pb2.py
"""
Placeholder protobuf message classes
These would be auto-generated from the .proto file
"""

class InterventionLevel:
    NONE = 0
    LEVEL_1_MONITOR = 1
    LEVEL_2_INTERVENTION = 2
    LEVEL_3_CRITICAL = 3

class InputType:
    TEXT = 0
    COMMAND = 1
    QUERY = 2
    CODE = 3

class EnergyLevel:
    HIGH = 0
    MEDIUM = 1
    LOW = 2
    EXHAUSTED = 3

class SleepStatus:
    WELL_RESTED = 0
    MILD_FATIGUE = 1
    MODERATE_FATIGUE = 2
    SEVERE_FATIGUE = 3

# Message classes
class InitializeRequest:
    def __init__(self, skill_id='', config=None):
        self.skill_id = skill_id
        self.config = config or {}

class InitializeResponse:
    def __init__(self, success=False, message='', session_id=''):
        self.success = success
        self.message = message
        self.session_id = session_id

class UserInput:
    def __init__(self, text='', timestamp='', type_val=0, metadata=None):
        self.text = text
        self.timestamp = timestamp
        self.type = type_val
        self.metadata = metadata or {}

class UserContext:
    def __init__(self, user_id_hash='', session_duration_minutes=0, recent_topics=None, energy_level=0, sleep_status=0):
        self.user_id_hash = user_id_hash
        self.session_duration_minutes = session_duration_minutes
        self.recent_topics = recent_topics or []
        self.energy_level = energy_level
        self.sleep_status = sleep_status

class AnalyzeInputRequest:
    def __init__(self, session_id='', user_input=None, user_context=None):
        self.session_id = session_id
        self.user_input = user_input or UserInput()
        self.user_context = user_context or UserContext()

class CognitiveState:
    def __init__(self, cognitive_load=0.0, stress_indicator=0.0, focus_level=0.0, active_biases=None):
        self.cognitive_load = cognitive_load
        self.stress_indicator = stress_indicator
        self.focus_level = focus_level
        self.active_biases = active_biases or []

class RiskPattern:
    def __init__(self, pattern_id='', pattern_name='', description='', severity=0, trigger_phrases=None):
        self.pattern_id = pattern_id
        self.pattern_name = pattern_name
        self.description = description
        self.severity = severity
        self.trigger_phrases = trigger_phrases or []

class AnalyzeInputResponse:
    def __init__(self, cognitive_state=None, detected_patterns=None, recommended_intervention=0, confidence_score=0.0):
        self.cognitive_state = cognitive_state or CognitiveState()
        self.detected_patterns = detected_patterns or []
        self.recommended_intervention = recommended_intervention
        self.confidence_score = confidence_score

class InterventionRequest:
    def __init__(self, session_id='', level=0, reason=''):
        self.session_id = session_id
        self.level = level
        self.reason = reason

class InterventionResponse:
    def __init__(self, intervention_applied=False, intervention_message='', suggested_actions=None):
        self.intervention_applied = intervention_applied
        self.intervention_message = intervention_message
        self.suggested_actions = suggested_actions or []

class StatusRequest:
    def __init__(self, session_id=''):
        self.session_id = session_id

class SystemMetrics:
    def __init__(self, cpu_usage=0.0, memory_usage=0.0, total_requests=0, average_response_time=0.0):
        self.cpu_usage = cpu_usage
        self.memory_usage = memory_usage
        self.total_requests = total_requests
        self.average_response_time = average_response_time

class StatusResponse:
    def __init__(self, skill_status='', active_sessions=0, metrics=None):
        self.skill_status = skill_status
        self.active_sessions = active_sessions
        self.metrics = metrics or SystemMetrics()

class PauseRequest:
    def __init__(self, session_id=''):
        self.session_id = session_id

class PauseResponse:
    def __init__(self, success=False, saved_state=''):
        self.success = success
        self.saved_state = saved_state

class ResumeRequest:
    def __init__(self, session_id='', saved_state=''):
        self.session_id = session_id
        self.saved_state = saved_state

class ResumeResponse:
    def __init__(self, success=False, message=''):
        self.success = success
        self.message = message

class DestroyRequest:
    def __init__(self, session_id=''):
        self.session_id = session_id

class DestroyResponse:
    def __init__(self, success=False, message=''):
        self.success = success
        self.message = message

class RunRequest:
    def __init__(self, session_id='', user_input=None):
        self.session_id = session_id
        self.user_input = user_input or UserInput()

class RunResponse:
    def __init__(self, response_id='', content='', intervention_level=0, requires_action=False):
        self.response_id = response_id
        self.content = content
        self.intervention_level = intervention_level
        self.requires_action = requires_action