# src/cognitive_shield_pb2_grpc.py
# Core protobuf definitions (auto-generated from proto/skill.proto)
# This file would be generated by the build process
# Included here for completeness

import grpc
from typing import Iterator

class CognitiveShieldSkillStub:
    def __init__(self, channel):
        self.Initialize = channel.unary_unary(
            '/skill_max_cognitive_shield.CognitiveShieldSkill/Initialize',
            request_serializer=None,
            response_deserializer=None,
        )
        self.AnalyzeInput = channel.unary_unary(
            '/skill_max_cognitive_shield.CognitiveShieldSkill/AnalyzeInput',
            request_serializer=None,
            response_deserializer=None,
        )
        # ... other methods

# Message classes would be auto-generated
class InitializeRequest:
    def __init__(self, skill_id='', config=None):
        self.skill_id = skill_id
        self.config = config or {}

class InitializeResponse:
    def __init__(self, success=False, message='', session_id=''):
        self.success = success
        self.message = message
        self.session_id = session_id

# ... other message classes

# gRPC service implementation
class CognitiveShieldSkillServicer:
    def Initialize(self, request, context):
        pass

    def AnalyzeInput(self, request, context):
        pass

    def RequestIntervention(self, request, context):
        pass

    def GetStatus(self, request, context):
        pass

    def Pause(self, request, context):
        pass

    def Resume(self, request, context):
        pass

    def Destroy(self, request, context):
        pass

    def Run(self, request, context):
        pass

def add_CognitiveShieldSkillServicer_to_server(servicer, server):
    rpc_method_handlers = {
        'Initialize': grpc.unary_unary_rpc_method_handler(
            servicer.Initialize,
            request_deserializer=InitializeRequest.from_bytes,
            response_serializer=InitializeResponse.SerializeToString,
        ),
        'AnalyzeInput': grpc.unary_unary_rpc_method_handler(
            servicer.AnalyzeInput,
            request_deserializer=None,
            response_serializer=None,
        ),
        # ... other methods
    }
    generic_handler = grpc.method_handlers_generic_handler(
        'skill_max_cognitive_shield.CognitiveShieldSkill', rpc_method_handlers)
    server.add_generic_rpcs(generic_handler)